moved squareToString to utils
